refactor(dictionary-evaluation): modularize evaluator internals into Helpers and SplitJoin domains
- move shared evaluator helpers into focused files:
  - DictionaryMatcher+EvaluationStylizedHelpers.swift
  - DictionaryMatcher+EvaluationSuffixHelpers.swift
  - DictionaryMatcher+EvaluationEvidenceHelpers.swift
- split split/join evaluator into focused files:
  - DictionaryMatcher+SplitJoinScoring.swift
  - DictionaryMatcher+SplitJoinForms.swift
  - DictionaryMatcher+SplitJoinGuards.swift
- keep strategy evaluators at evaluation root:
  - DictionaryMatcher+StandardEvaluation.swift
  - DictionaryMatcher+MergedTokenEvaluation.swift
  - DictionaryMatcher+ThreeTokenEvaluation.swift
- remove obsolete files:
  - DictionaryMatcher+EvaluationShared.swift
  - DictionaryMatcher+SplitJoinEvaluation.swift
- update Xcode project file membership and group structure for new folders/files
- update Docs/CODEMAP.md and Docs/ENGINEERING.md to reflect the new architecture
- preserve runtime behavior (pure refactor / maintainability-focused change)
